+++ crates/module-system/hyperlane/tests/integration/with_agent/mod.rs
@@ -133,7 +133,7 @@ async fn test_relayer_basic_dispatch_process() {
     submit_tx(&rollup.api_client, dispatch_tx).await;
 
     // look for `process` event
-    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 10 {
+    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 15 {
         let events = next_slot_events(&rollup.api_client, &mut slot_subscription).await;
 
         if let Some(process_event) = find_event(&events, "Mailbox/Process") {
@@ -211,7 +211,7 @@ async fn test_multisig_ism() {
     submit_tx(&rollup.api_client, dispatch_tx).await;
 
     // look for `process` event
-    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 10 {
+    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 15 {
         let events = next_slot_events(&rollup.api_client, &mut slot_subscription).await;
 
         if let Some(process_event) = find_event(&events, "Mailbox/Process") {
@@ -281,7 +281,7 @@ async fn test_process_message_from_evm_counterparty() {
     hyperlane.mine_next_block_on_counterparty().await;
 
     // look for `process` event
-    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 10 {
+    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 15 {
         let events = next_slot_events(&rollup.api_client, &mut slot_subscription).await;
 
         if let Some(process_event) = find_event(&events, "Mailbox/Process") {
@@ -350,7 +350,7 @@ async fn test_dispatch_message_to_evm_counterparty() {
     submit_tx(&rollup.api_client, dispatch_tx).await;
 
     // look for `dispatch` event
-    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 10 {
+    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 15 {
         let events = next_slot_events(&rollup.api_client, &mut slot_subscription).await;
 
         if let Some(process_event) = find_event(&events, "Mailbox/Dispatch") {
@@ -449,7 +449,7 @@ async fn test_warp_transfer_back_and_forth_with_evm_counterparty(
     let mut local_route_id = HexString([0; 32]);
     let mut remote_route_id = HexString([0; 32]);
     // look for `route registered` event
-    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 10 {
+    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 15 {
         let events = next_slot_events(&rollup.api_client, &mut slot_subscription).await;
 
         if let Some(route_registered_event) = find_event(&events, "Warp/RouteRegistered") {
@@ -497,7 +497,7 @@ async fn test_warp_transfer_back_and_forth_with_evm_counterparty(
     hyperlane.mine_next_block_on_counterparty().await;
 
     let mut transfer_received = false;
-    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 10 {
+    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 15 {
         let events = next_slot_events(&rollup.api_client, &mut slot_subscription).await;
 
         if let Some(token_recv_event) = find_event(&events, "Warp/TokenTransferReceived") {
@@ -550,7 +550,7 @@ async fn test_warp_transfer_back_and_forth_with_evm_counterparty(
     let transfer_tx = encode_call(prover.user_info.private_key(), &transfer_call);
     submit_tx(&rollup.api_client, transfer_tx).await;
 
-    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 10 {
+    for _ in 0..DEFAULT_FINALIZATION_BLOCKS * 15 {
         let events = next_slot_events(&rollup.api_client, &mut slot_subscription).await;
 
         // look for event that sent outbound transfer